Query Validation Issue
I noticed that if you try to validate a language agnostic query, you receive the message "languages cannot be null (missing)".

As a current counter measure I think it'd be good to just validate against any language (I've chosen english in this case).

Upon doing so, the query validates, and the query uploads as language agnostic.